phase-5/backend/requirements.txt
# Phase 5 Backend Requirements
# Last Updated: 2026-02-04

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Dapr Integration
dapr==1.12.0
dapr-devtools==0.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML (Ollama integration)
ollama==0.1.6
openai==1.10.0  # For API-compatible fallback

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
testcontainers==4.5.1
testcontainers-kafka==4.5.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
orjson==3.9.12
